[move]type to @/types
